backend/pyproject.toml
[project]
name = "kb"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.128.0",
    "uvicorn[standard]>=0.34",
    "python-multipart>=0.0.20",
    "sse-starlette>=2.3",
    "litellm==1.83.3",
    "pydantic-settings>=2.9",
    "PyJWT>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4",
    "pytest-asyncio>=0.26",
    "httpx>=0.28",
    "pytest-mock>=3.14",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
